Updating Instagram Platform API postman collection with IGv3 Send API
Summary:
Updating the "Instagram Platform API" postman collection with the IGv3 Send API

The previous postman collection that was checked in only had the "Send API". We already had "Instagram Platform API" collection created with pre-existing APIs in it. So after discussing the options with Matt and Sandeep, we decied to merge the Send API into this existing collection.

Here's what I did
1. Downloaded this postman collection: https://www.postman.com/meta/instagram/collection/6yqw8pt/instagram-api
2. Merged my Send API into this one
3. Exported new json file
4. Create this diff

Reviewed By: mattsdni

Differential Revision: D74196759

fbshipit-source-id: 345e4acd27d3ba41283f8ffefdba2dd6b070e4e7